requirements.txt
beautifulsoup4
PrettyTables
json
requests
